fixes: build and script ref in HTML
Removed "integrity" attribute in ArcGIS API script tag. Esri changed the
file which broke the integrity tag.

Demo build: CSS files are now copied in the "prepare" npm script.
Prior to this, when running the demo page the CSS files would be missing
and thus the control would not work correctly.